Fixed subtle bug in section marking code.
Thanks to Linkvitch for helping find this one! Also, found and removed
some more 'GPU in Main' cruft hiding out in sect.c. Gone!

+++ mark.h
@@ -13,7 +13,7 @@
 #include "sect.h"
 
 #define MARK_ALLOC_INCR 1024		// # bytes to alloc for more mark space 
-#define MIN_MARK_MEM    (3*sizeof(WORD)+2*sizeof(LONG))
+#define MIN_MARK_MEM    (3 * sizeof(WORD) + 2 * sizeof(LONG))
 
 // Globals, Externals etc
 extern MCHUNK * firstmch;
@@ -21,7 +21,8 @@ extern MCHUNK * firstmch;
 // Prototypes
 void init_mark(void);
 void stopmark(void);
-int rmark(int, LONG, int, int, SYM *);
+//int rmark(int, LONG, int, int, SYM *);
+int rmark(uint16_t, uint32_t, uint16_t, uint16_t, SYM *);
 int amark(void);
 LONG bsdmarkimg(char *, LONG, LONG, int);
